MaxFloatNumAmongThreeNum

+++ Generics_MaximumNumber.java
@@ -1,12 +1,11 @@
 package com.bl.program;
-
 import java.util.Scanner;
 
 public class TestMaximum  {
 	Scanner sc = new Scanner(System.in);
 	
-	public Integer getMaxNumber(Integer num1, Integer num2, Integer num3) {
-		Integer max = num1;
+	public Double getMaxNumber(Double num1, Double num2, Double num3) {
+		Double max = num1;
 		if(num2.compareTo(max) > 0) 
 			max = num2;
 		else if(num3.compareTo(max) > 0)
